safari-534.53-branch/LayoutTests/accessibility/editable-webarea-context-menu-point-expected.txt

This tests that the click point, used for opening context menus, changes based on the current selection in an editable web area. This tests that the click point, used for opening context menus, changes based on the current selection in an editable web area.
This tests that the click point, used for opening context menus, changes based on the current selection in an editable web area (by setting the selection twice and comparing the click point locations)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
